Infers mesh shape when using state builder.
